Move functions from bids_utils to oasis_to_bids_utils + create test script
