+++ test/unittests/iotools/converters/oasis_to_bids/test_oasis_to_bids_utils.py
@@ -1,0 +1,10 @@
+import pytest
+
+from clinica.iotools.converters.oasis_to_bids.oasis_to_bids_utils import (
+    create_sessions_dict,
+)
+
+
+def test_create_sessions_dict():
+    # todo
+    pass
